mcp_clients_py/basic_client/pyproject.toml
[project]
name = "basic-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.9.2",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
]
